Split BossBarManager into multiple classes for clarity
